requirements.txt
# Core dependencies for AI agent
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20

# MCP (Model Context Protocol) dependencies
mcp>=0.1.0

# Research source integrations
arxiv>=2.0.0
biopython>=1.83

# HTTP and data handling
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
